Restructure format
Formatting was messed up from my misinterpretation of the code, should be fixed now
