+++ arith/array2/.gitignore
