+++ arith/array2/src/lib.rs
