+++ arith/rpeg/src/codec.rs
@@ -1,0 +1,9 @@
+use csc411_image;
+
+pub fn compress(filename: Option<&str>) {
+    todo!();
+}
+
+pub fn decompress(filename: Option<&str>) {
+    todo!();
+}
