+++ arith/rpeg/src/lib.rs
@@ -1,0 +1,27 @@
+pub fn prepare_ppm() -> todo!() {
+
+}
+
+pub fn convert_to_rgb() -> todo!() {
+
+}
+
+pub fn convert_to_component_video() -> todo!() {
+
+}
+
+pub fn write_to_standard_output() -> todo!() {
+
+}
+
+pub fn convert_to_four_bit() -> todo!() {
+
+}
+
+pub fn discrete_cosine_transfer() -> todo!() {
+
+}
+
+pub fn inverse_discrete_cosine_transfer() -> todo!() {
+
+}
