+++ arith/rpeg/src/main.rs
@@ -1,7 +1,19 @@
 use csc411_image;
 use csc411_arith;
-
+use std::env;
+use rpeg::codec::{compress, decompress};
 
 fn main() {
-    println!("Hello, world!");
+    let args: Vec<String> = env::args().collect();
+    let argnum = args.len();
+    assert!(argnum == 2 || argnum == 3);
+    let filename = args.iter().nth(2).unwrap();
+    match args[1].as_str() {
+        "-c" => compress(Some(filename)),
+        "-d" => decompress(Some(filename)),
+        _ => {
+            eprintln!("Usage: rpeg -d [filename]\nrpeg -c [filename]")
+        }
+    }
 }
+
