--- bitpack/src/codec.rs
@@ -1,9 +1,0 @@
-use csc411_image;
-
-pub fn compress(filename: Option<&str>) {
-    todo!();
-}
-
-pub fn decompress(filename: Option<&str>) {
-    todo!();
-}
